ChannelSender: Use `Arc` to ensure that the fuse device isn't closed
...before the users have finished with it.

The previous implementation was dangerous.  Closing the FD doesn't
guarantee that users will be woken up or that subsequent writes would
return EBADF.  Instead the FD could be reused and our implementation would
end up writing to an unexpected destination.  This way is much safer.

Users will be reliably interrupted when the filesystem is unmounted
however.

`ChannelSender` is now no-longer `Copy`, but is instead `Clone`.  I don't
think this is an API break as `ChannelSender` isn't a public API, except
for appearing in the `Session` constructor.