ll: Make Request a trait, rather than a struct
The struct is replaced by `AnyRequest`.  All our operation types now
implement `Request`, in doing so they're a bit more self-contained.

There were a few clues that this was the better design when I was working
on the previous design.  There were a few places where I was having to
explicitly extract data from the header and inject it when constructing
these types.  Now all the information is injected everywhere.

This is a precursor to adding `reply` members to these structs.

We can do more with this later too - like applying `FilenameInDir` to all
the structs that have a filename.  This might allow for some useful generic
patterns.